Move CheckCall to openblas_build::
